requirements.txt
django==5.1.1
python-dotenv==1.0.1
djangorestframework==3.15.2
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.9
django-filter==24.3
drf-yasg==1.21.7





